requirements.txt
altair==6.0.0
attrs==25.4.0
black==25.11.0
bleach==6.3.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.0
fonttools==4.61.0
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kaggle==1.8.2
kagglesdk==0.1.13
kiwisolver==1.4.9
librt==0.7.3
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
mypy==1.19.0
mypy_extensions==1.1.0
narwhals==2.13.0
networkx==3.6
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.1
protobuf==6.33.2
pyarrow==22.0.0
pydeck==0.9.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytokens==0.3.0
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
smmap==5.0.2
streamlit==1.52.1
sympy==1.14.0
tenacity==9.1.2
text-unidecode==1.3
threadpoolctl==3.6.0
toml==0.10.2
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tornado==6.5.2
tqdm==4.67.1
types-requests==2.32.4.20250913
types-tqdm==4.67.0.20250809
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.0
webencodings==0.5.1
